[chore][pkg/stanza] Cleanup udp input operator files
